Rename `MarkDown` -> `Markdown` to resolve differences for non-Windows machines
