--- src/modules/previewpane/MarkDownPreviewHandler/Telemetry/Events/MarkdownFilePreviewError.cs
@@ -1,23 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation
-// The Microsoft Corporation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.PowerToys.Telemetry;
-using Microsoft.PowerToys.Telemetry.Events;
-
-namespace MarkdownPreviewHandler.Telemetry.Events
-{
-    /// <summary>
-    /// A telemetry event that is triggered when an error occurs while attempting to view a markdown file in the preview pane.
-    /// </summary>
-    public class MarkdownFilePreviewError : EventBase, IEvent
-    {
-        /// <summary>
-        /// Gets or sets the error message.
-        /// </summary>
-        public string Message { get; set; }
-
-        /// <inheritdoc/>
-        public PartA_PrivTags PartA_PrivTags => PartA_PrivTags.ProductAndServicePerformance;
-    }
-}
