+++ src/modules/previewpane/MarkdownPreviewHandler/MarkdownPreviewHandler.cs
