+++ src/modules/previewpane/MarkdownPreviewHandler/MarkdownPreviewHandlerControl.cs
