+++ src/modules/previewpane/MarkdownPreviewHandler/Properties/Settings.settings
